+++ tectonic/dpx-dpxfile.c
@@ -425,7 +425,7 @@ dpx_create_temp_file (void)
 
 static int
 dpx_clear_cache_filter (const struct dirent *ent) {
-    int plen = strlen(DPX_PREFIX);
+    size_t plen = strlen(DPX_PREFIX);
     if (strlen(ent->d_name) != plen + MAX_KEY_LEN * 2) return 0;
     return strstartswith(ent->d_name, DPX_PREFIX) == 0;
 }
